Removing some redundant code
